home, footer, menu, adaptive

+++ template-parts/content_home.php
@@ -9,7 +9,7 @@
 
 ?>
 
-<section class="slider pt-5 pb-4">
+<section class="slider pb-4">
 	<div class="container">
 		<div class="top_slider">
 			<?php 
@@ -26,7 +26,9 @@
 						<div class="slide position-relative">
 							<div class="slide_image">
 								<?php
+									echo '<a href="', the_permalink(), '">';
 									the_post_thumbnail( 'full' );
+									echo '</a>';
 								?>
 							</div>
 							<div class="slide_description bg-white p-4">
@@ -36,9 +38,9 @@
 								<div class="post_info">
 									<?php
 									if ( 'post' === get_post_type() ) : ?>
-										<?php printf( __( '<span class="fn">%s:</span>' ), get_the_author_link() ); ?>
-										<span class="date">
-											<?php printf( get_the_date() ); ?>
+										<?php printf( __( 'by <span class="author">%s</span>' ), get_the_author_link() ); ?>
+										<span class="date text-muted">
+											at <?php printf( get_the_date() ); ?>
 										</span>
 									<?php
 									endif; 
@@ -65,14 +67,17 @@
 				<div class="row">
 					<?php
 						foreach( get_categories(['hide_empty' => false]) as $category) {
-						    echo '<div class="col-6 category_block">
+							if ( function_exists('get_wp_term_image') ) {
+								echo '<div class="col-6 category_block">
 									<div class="category_image">';
-							    
-							echo do_shortcode(sprintf('[wp_custom_image_category term_id="%s"]',$category->term_id)) ;
-							echo '</div>
-								  <div class="category_description">';
-							echo '<h3><a class="btn btn-primary text-white" href="' . get_category_link( $category->term_id ) . '" title="' . sprintf( __( "View all posts in %s" ), $category->name ) . '" ' . '>' . $category->name . '</a></h3>';
-							echo '</div></div>' ;
+									
+								echo '<img src="' .  get_wp_term_image($category->term_id) . '">';
+
+								echo '</div>
+									<div class="category_description">';
+								echo '<h3><a class="btn btn-primary text-white" href="' . get_category_link( $category->term_id ) . '" title="' . sprintf( __( "View all posts in %s" ), $category->name ) . '" ' . '>' . $category->name . '</a></h3>';
+								echo '</div></div>';
+							};
 						}
 					?>
 				</div>
@@ -89,17 +94,24 @@
 					$args = array( 'numberposts' => '6' );
 					$recent_posts = wp_get_recent_posts( $args );
 					foreach( $recent_posts as $recent ){
-						echo '<div class="card">';
-						if ( has_post_thumbnail( $recent["ID"]) ) {
-							echo  get_the_post_thumbnail($recent["ID"],'medium');
-						}
+						echo '<div class="card card_slider">';
+							if ( has_post_thumbnail( $recent["ID"]) ) {
+								echo '<div class="card-image card-image">';
+								echo  '<a href="' . get_permalink($recent["ID"]) . '">' . get_the_post_thumbnail( $recent["ID"], 'tumbnail') . '</a>';
+								echo '</div>';
+							}
 						echo '<div class="card-body">';
-						echo '<h5 class="card-title"><a href="' . get_permalink($recent["ID"]) . '">' .   $recent["post_title"].'</a> </h3> ';
-        				echo '<p class="card-text">' . my_excerpt( array( 'text' => $recent['post_content'], 'maxchar' => 100) ) . '</p>';
-
-        				echo '<a class="btn btn-primary text-white" href="' . get_permalink($recent["ID"]) . '">Читать</a>';
+						echo '<div class="card-data">
+								<span class="text-muted">' . date('n.j.Y', strtotime($recent['post_date'])) .'</span>';
+								?>
+								<?php
+									printf( __( 'by <span class="author">%s</span>' ), get_the_author_link() ); 
+									echo '<span class="ml-3 comments"><i class="fas fa-comment mr-2"></i>' . get_comments_number($recent["ID"]) . '</span>';
+								?>
+							 <?
         				echo '</div>';
-        				echo '</div>';
+						echo '<h5 class="card-title"><a href="' . get_permalink($recent["ID"]) . '">' .   $recent["post_title"].'</a></h3> ';
+        				echo '</div></div>';
 					}
 				?>
 			</div>
